move utf convertion to utf.h
