+++ src/common.h
@@ -25,6 +25,8 @@
 
 #include "utils/log.h"
 #include "utils/nums.h"
+#include "utils/string_utils.h"
+#include "utils/utf.h"
 
 using namespace std;
 
@@ -70,321 +72,6 @@ inline int indexOf(const vector<T>& v, const T& x) {
     auto it = find(v.begin(), v.end(), x);
     int i = std::distance(v.begin(), it);
     return (i >= v.size() ? -1 : i);
-}
-
-/**
- * Convert a value to string
- */
-template <class T>
-inline string tostring(T val) {
-    ostringstream os;
-    os << val;
-    return os.str();
-}
-
-/**
- * Convert a value to wide string
- */
-template <class T>
-inline wstring towstring(T val) {
-    wostringstream os;
-    os << val;
-    return os.str();
-}
-
-template <class T>
-inline void valueof(const string& s, T& val) {
-    stringstream ss;
-    ss << s;
-    ss >> val;
-}
-
-template <class T>
-inline void valueof(const wstring& s, T& val) {
-    wstringstream ss;
-    ss << s;
-    ss >> val;
-}
-
-inline bool str2int(const string& str, int& res, int radix) {
-    char* endptr = nullptr;
-    errno = 0;
-
-    const long val = strtol(str.c_str(), &endptr, radix);
-
-    if ((val == LONG_MAX || val == LONG_MIN) && errno == ERANGE)
-        return false;
-
-    res = static_cast<int>(val);
-    return endptr == str.c_str() + str.size();
-}
-
-/**
- * Transform a string to lowercase
- */
-inline string& tolower(_out_ string& src) {
-    transform(src.begin(), src.end(), src.begin(), ::tolower);
-    return src;
-}
-
-inline wstring& tolower(_out_ wstring& src) {
-    transform(src.begin(), src.end(), src.begin(), ::tolower);
-    return src;
-}
-
-/**
- * Ignore left side whitespace in a string
- */
-inline string& ltrim(string& s) {
-    s.erase(s.begin(), find_if(s.begin(), s.end(), not1(ptr_fun<int, int>(isspace))));
-    return s;
-}
-
-/**
- * Ignore right side whitespace in a string
- */
-inline string& rtrim(string& s) {
-    s.erase(find_if(s.rbegin(), s.rend(), not1(ptr_fun<int, int>(isspace))).base(), s.end());
-    return s;
-}
-
-/**
- * Ignore left and right side whitespace in a string
- */
-inline string& trim(string& s) {
-    return ltrim(rtrim(s));
-}
-
-/**
- * Split string with specified delimeter
- */
-inline void split(const string& str, char del, _out_ vector<string>& res) {
-    stringstream ss(str);
-    string tok;
-    while (getline(ss, tok, del)) res.push_back(tok);
-}
-
-inline bool startswith(const string& str, const string& cmp) {
-    return str.find(cmp) == 0;
-}
-
-inline bool endswith(const string& str, const string& cmp) {
-    return str.rfind(cmp) == (str.length() - cmp.length());
-}
-
-inline bool startswith(const wstring& str, const wstring& cmp) {
-    return str.find(cmp) == 0;
-}
-
-inline bool endswith(const wstring& str, const wstring& cmp) {
-    return str.rfind(cmp) == (str.length() - cmp.length());
-}
-
-/**
- * Split string with delimiter
- */
-class strtokenizer {
-private:
-    string _str;
-    string _del;
-    bool _ret;
-    int _pos;
-
-public:
-    strtokenizer(const string& str) {
-        _str = str;
-        _del = " \t\n\r\f";
-        _ret = false;
-        _pos = 0;
-    }
-
-    strtokenizer(const string& str, const string& del, bool ret = false) {
-        _str = str;
-        _del = del;
-        _ret = ret;
-        _pos = 0;
-    }
-
-    int count_tokens() {
-        int c = 0;
-        bool in = false;
-        for (int i = _pos, len = _str.length(); i < len; i++) {
-            if (_del.find(_str[i]) != string::npos) {
-                if (_ret) c++;
-                if (in) {
-                    c++;
-                    in = false;
-                }
-            } else {
-                in = true;
-            }
-        }
-        if (in) c++;
-        return c;
-    }
-
-    string next_token() {
-        int i = _pos;
-        int len = _str.length();
-
-        if (i < len) {
-            if (_ret) {
-                if (_del.find(_str[_pos]) != string::npos)
-                    return string({_str[_pos++]});
-                for (_pos++; _pos < len; _pos++)
-                    if (_del.find(_str[_pos]) != string::npos)
-                        return _str.substr(i, _pos - i);
-                return _str.substr(i);
-            }
-
-            while (i < len && _del.find(_str[i]) != string::npos) i++;
-
-            _pos = i;
-            if (i < len) {
-                for (_pos++; _pos < len; _pos++)
-                    if (_del.find(_str[_pos]) != string::npos)
-                        return _str.substr(i, _pos - i);
-                return _str.substr(i);
-            }
-        }
-        return "";
-    }
-};
-
-/**
- * Convert unicode wide string to UTF-8 encoded string
- */
-inline string& wide2utf8(const wchar_t* in, _out_ string& out) {
-    unsigned int codepoint = 0;
-    for (; *in != 0; ++in) {
-        if (*in >= 0xd800 && *in <= 0xdbff) {
-            codepoint = ((*in - 0xd800) << 10) + 0x10000;
-        } else {
-            if (*in >= 0xdc00 && *in <= 0xdfff) {
-                codepoint |= *in - 0xdc00;
-            } else {
-                codepoint = *in;
-            }
-
-            if (codepoint <= 0x7f) {
-                out.append(1, static_cast<char>(codepoint));
-            } else if (codepoint <= 0x7ff) {
-                out.append(1, static_cast<char>(0xc0 | ((codepoint >> 6) & 0x1f)));
-                out.append(1, static_cast<char>(0x80 | (codepoint & 0x3f)));
-            } else if (codepoint <= 0xffff) {
-                out.append(1, static_cast<char>(0xe0 | ((codepoint >> 12) & 0x0f)));
-                out.append(1, static_cast<char>(0x80 | ((codepoint >> 6) & 0x3f)));
-                out.append(1, static_cast<char>(0x80 | (codepoint & 0x3f)));
-            } else {
-                out.append(1, static_cast<char>(0xf0 | ((codepoint >> 18) & 0x07)));
-                out.append(1, static_cast<char>(0x80 | ((codepoint >> 12) & 0x3f)));
-                out.append(1, static_cast<char>(0x80 | ((codepoint >> 6) & 0x3f)));
-                out.append(1, static_cast<char>(0x80 | (codepoint & 0x3f)));
-            }
-            codepoint = 0;
-        }
-    }
-    return out;
-}
-
-inline string wide2utf8(const wchar_t* in) {
-    string out;
-    return wide2utf8(in, out);
-}
-
-/**
- * Convert an UTF-8 encoded char sequence to wide unicode string,
- * the encoding of input char sequence must be known as UTF-8
- */
-inline wstring& utf82wide(const char* in, _out_ wstring& out) {
-    if (in == nullptr) return out;
-    unsigned int codepoint;
-    while (*in != 0) {
-        unsigned char ch = static_cast<unsigned char>(*in);
-        if (ch <= 0x7f)
-            codepoint = ch;
-        else if (ch <= 0xbf)
-            codepoint = (codepoint << 6) | (ch & 0x3f);
-        else if (ch <= 0xdf)
-            codepoint = ch & 0x1f;
-        else if (ch <= 0xef)
-            codepoint = ch & 0x0f;
-        else
-            codepoint = ch & 0x07;
-        ++in;
-        if (((*in & 0xc0) != 0x80) && (codepoint <= 0x10ffff)) {
-            if (codepoint > 0xffff) {
-                out.append(1, static_cast<wchar_t>(0xd800 + (codepoint >> 10)));
-                out.append(1, static_cast<wchar_t>(0xdc00 + (codepoint & 0x03ff)));
-            } else if (codepoint < 0xd800 || codepoint >= 0xe000)
-                out.append(1, static_cast<wchar_t>(codepoint));
-        }
-    }
-    return out;
-}
-
-inline wstring utf82wide(const char* in) {
-    wstring out;
-    return utf82wide(in, out);
-}
-
-/**
- * Returns a replacement string for the given one that has all backslashes
- * and dollar signs escaped
- */
-inline string& quotereplace(const string& src, _out_ string& out) {
-    for (size_t i = 0; i < src.length(); i++) {
-        char c = src[i];
-        if (c == '\\' || c == '$') out.append(1, '\\');
-        out.append(1, c);
-    }
-    return out;
-}
-
-inline wstring& quotereplace(const wstring& src, _out_ wstring& out) {
-    for (size_t i = 0; i < src.length(); i++) {
-        wchar_t c = src[i];
-        if (c == L'\\' || c == L'$') out.append(1, L'\\');
-        out.append(1, c);
-    }
-    return out;
-}
-
-/**
- * Replace string with specified string in the first
- */
-inline string& replacefirst(_out_ string& src, const string& from, const string& to) {
-    size_t start = src.find(from);
-    if (start == string::npos) return src;
-    src.replace(start, from.length(), to);
-    return src;
-}
-
-inline string& replaceall(_out_ string& src, const string& from, const string& to) {
-    if (from.empty()) return src;
-    size_t start = 0;
-    while ((start = src.find(from, start)) != string::npos) {
-        src.replace(start, from.length(), to);
-        start += to.length();
-    }
-    return src;
-}
-
-inline wstring& replacefirst(_out_ wstring& src, const wstring& from, const wstring& to) {
-    size_t start = src.find(from);
-    if (start == wstring::npos) return src;
-    src.replace(start, from.length(), to);
-    return src;
-}
-
-inline wstring& replaceall(_out_ wstring& src, const wstring& from, const wstring& to) {
-    if (from.empty()) return src;
-    size_t start = 0;
-    while ((start = src.find(from, start)) != wstring::npos) {
-        src.replace(start, from.length(), to);
-        start += to.length();
-    }
-    return src;
 }
 
 /***************************************************************************************************
